feat: Added a 'ready' check.
The check ensures that at least 80% of the pieces are accounted for. The reasoning for 80% is documented, however the fact that we have to run this check may point to a lack of understanding of bitfields on our side or an issue with what peers we're receiving. More research is needed.