fix: Remove thread panics when peer connection refuses
Sometimes, peers just don't connect to use at all (due to the nature of
the internet), and currently, it will crash our entire thread if a peer
fails to connect due to our careless usage of unwrap. This PR intends to
fix that.

This is done by changing `PeerStream::connect` into a `Result` instead
of just unwrapping it (no clue why we didn't do this originally).